Update for new database scheme

+++ db/g_setting.php
@@ -24,18 +24,8 @@
         $GLOBALS['responseMsg'] = $GLOBALS['responseMsg'] . $msg;
     }
 
-    // 檢查群組是否存在
-    $sql = sprintf("SELECT GroupID FROM groupinfo WHERE GroupID = '%s'"
-            , mysql_real_escape_string($groupID));
-    $result = mysql_query($sql) or responseMsg('Invalid query #1: ' . mysql_error() . "\n");
-
-    if (mysql_num_rows($result) == 0) {
-        // 新增群組
-        $sql = sprintf("INSERT INTO groupinfo (GroupID, GroupName) VALUES ('%s', '%s')"
-                , mysql_real_escape_string($groupID)
-                , mysql_real_escape_string($groupNAME));
-        $result = mysql_query($sql) or responseMsg('Invalid query #2: ' . mysql_error() . "\n");
-        responseMsg(( $result ? $sql."\n" : '' ));
+    // 儲存資料
+    function store( $data ){
         // 儲存管理員資料
         foreach ( $data->admins as $admins) {
             $sql = sprintf("INSERT INTO userinfo VALUES ('%s', '%s')"
@@ -44,7 +34,7 @@
             mysql_query($sql) or responseMsg('Invalid query #3: ' . mysql_error() . "\n");
             $sql = sprintf("INSERT INTO belongsto VALUES ('%s', '%s', '%s')"
                 , mysql_real_escape_string($admins->id)
-                , mysql_real_escape_string($groupID)
+                , mysql_real_escape_string($data->ID)
                 , mysql_real_escape_string('ADMIN'));
             $result = mysql_query($sql) or responseMsg('Invalid query #4: ' . mysql_error() . "\n");
             responseMsg(( $result ? $sql."\n" : '' ));
@@ -57,14 +47,30 @@
             mysql_query($sql) or responseMsg('Invalid query #5: ' . mysql_error() . "\n");
             $sql = sprintf("INSERT INTO belongsto VALUES ('%s', '%s', '%s')"
                 , mysql_real_escape_string($members->id)
-                , mysql_real_escape_string($groupID)
-                , mysql_real_escape_string('MEMBERS'));
+                , mysql_real_escape_string($data->ID)
+                , mysql_real_escape_string('MEMBER'));
             $result = mysql_query($sql) or responseMsg('Invalid query #6: ' . mysql_error() . "\n");
             responseMsg(( $result ? $sql."\n" : '' ));
         }
+    }
+
+    // 檢查群組是否存在
+    $sql = sprintf("SELECT GroupID FROM groupinfo WHERE GroupID = '%s'"
+            , mysql_real_escape_string($groupID));
+    $result = mysql_query($sql) or responseMsg('Invalid query #1: ' . mysql_error() . "\n");
+
+    if (mysql_num_rows($result) == 0) {
+        // 新增群組
+        $sql = sprintf("INSERT INTO groupinfo (GroupID, GroupName) VALUES ('%s', '%s')"
+                , mysql_real_escape_string($groupID)
+                , mysql_real_escape_string($groupNAME));
+        $result = mysql_query($sql) or responseMsg('Invalid query #2: ' . mysql_error() . "\n");
+        responseMsg(( $result ? $sql."\n" : '' ));
+
+        store($data);
 
     } else {
-        if ( isset( $data->admins ) ) {
+        if ( isset( $data->admins ) && isset( $data->members ) ) {
             // 需要更新 belongsto
             // 清空舊資料
             $sql = sprintf("DELETE FROM belongsto WHERE GroupID = '%s'"
@@ -72,24 +78,7 @@
             mysql_query($sql) or responseMsg('Invalid query #7: ' . mysql_error() . "\n");
             responseMsg($sql."\n");
 
-            // 更新管理員資料
-            foreach ( $data->admins as $admins) {
-                $sql = sprintf("INSERT INTO belongsto VALUES ('%s', '%s', '%s')"
-                    , mysql_real_escape_string($admins->id)
-                    , mysql_real_escape_string($groupID)
-                    , mysql_real_escape_string('ADMIN'));
-                $result = mysql_query($sql) or responseMsg('Invalid query #8: ' . mysql_error() . "\n");
-                responseMsg(( $result ? $sql."\n" : '' ));
-            }
-            // 更新成員資料
-            foreach ( $data->members as $members) { 
-                $sql = sprintf("INSERT INTO belongsto VALUES ('%s', '%s', '%s')"
-                    , mysql_real_escape_string($members->id)
-                    , mysql_real_escape_string($groupID)
-                    , mysql_real_escape_string('MEMBERS'));
-                $result = mysql_query($sql) or responseMsg('Invalid query #9: ' . mysql_error() . "\n");
-                responseMsg(( $result ? $sql."\n" : '' ));
-            }
+            store($data);
 
         } else {
             // 更新群組資料
